requirements.txt
streamlit
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pillow
streamlit-paste-button
folium==0.17.0
streamlit-folium==0.22.0
geopy==2.4.1
requests==2.32.3
